+++ src/cli/project.py
@@ -1,6 +1,5 @@
 from cmd import Cmd
 import project.project as project
-import context.context as context
 
 
 class ProjectMenu(Cmd):
@@ -20,9 +19,10 @@ class ProjectMenu(Cmd):
 
         super().default(line)
 
-    def make_context(self) -> context.Context:
-        ctx = context.Context()
-        ctx.set_project(self.proj)
+    def make_context(self) -> dict:
+        ctx = {
+            'project': self.proj
+        }
         return ctx
 
     def do_q(self, args):