+++ src/context/context.py
@@ -1,9 +1,0 @@
-# import project.project as project
-
-
-class Context:
-    proj: str
-
-    def set_project(self, proj: str):
-        self.proj = proj
-